Export resource files from evernote

+++ ever2simple/converter.py
@@ -6,14 +6,19 @@ from csv import DictWriter
 from dateutil.parser import parse
 from html2text import HTML2Text
 from lxml import etree
+import re
+import unicodedata
+import StringIO
+import hashlib
+
 
 
 class EverConverter(object):
     """Evernote conversion runner
     """
 
-    fieldnames = ['createdate', 'modifydate', 'content', 'tags', 'title']
-    date_fmt = '%h %d %Y %H:%M:%S'
+    fieldnames = ['createdate', 'modifydate', 'content', 'tags', 'resources']
+    date_fmt = '%Y %m %d %H:%M:%S'
 
     def __init__(self, enex_filename, simple_filename, fmt, metadata):
         self.enex_filename = os.path.expanduser(enex_filename)
@@ -39,10 +44,67 @@ class EverConverter(object):
     def prepare_notes(self, xml_tree):
         notes = []
         raw_notes = xml_tree.xpath('//note')
+        note_ids = {}
         for note in raw_notes:
             note_dict = {}
             title = note.xpath('title')[0].text
             note_dict['title'] = title
+            # Check for duplicates
+            # Filename is truncated to 100 characters (Windows MAX_PATH is 255) - not a perfect fix but an improvement
+
+            orig_nfilename = self._format_filename(title[0:99])
+            nfilename = orig_nfilename
+            count = 0
+            while (nfilename in note_ids):
+                count = count + 1
+                nfilename = orig_nfilename + "(" + str(count) + ")"
+            note_dict["filename"] = nfilename
+            note_ids[nfilename] = note_dict
+
+            resources = []
+            resources_by_id = {}
+            resources_by_fname = {}
+            for resource in note.xpath("resource"):
+                mime = resource.xpath("mime")[0].text
+                tag = unicode("""[{}]({})""")
+                res_id_rx = re.compile("""objID="(\w+)" """)
+                ext_rx = re.compile("""\w+/(\w+)""")
+
+                try:
+                    recogn = resource.xpath("recognition")[0].text
+                    res_id = res_id_rx.findall(recogn)[0]
+                except IndexError:
+                    m = hashlib.md5()
+                    m.update(resource.xpath("data")[0].text.decode("base64"))
+                    res_id = m.hexdigest()
+                    
+
+                r_title = res_id + "." + ext_rx.findall(mime)[0]
+
+                if mime.startswith("image"):
+                    tag = unicode("""![]({})""")
+
+                orig_name = None
+                try:
+                    orig_name = resource.xpath("resource-attributes")[0].xpath("file-name")[0].text
+                except IndexError:
+                    pass
+
+                if orig_name is not None and len(orig_name) > 0:
+                    rcount = 0
+                    r_title = orig_name
+                    while (r_title in resources_by_fname):
+                        rcount = rcount + 1
+                        r_title = "(" + str(rcount) + ")" + orig_name 
+
+                rfilename = nfilename + "_" + self._format_filename(r_title)
+                data = resource.xpath("data")[0].text
+                res = {"filename": rfilename, "data": data, "tag": tag, "used": 0}
+                resources.append(res)
+                resources_by_id[res_id] = res
+                resources_by_fname[r_title] = res
+            note_dict['resources'] = resources
+
             # Use dateutil to figure out these dates
             # 20110610T182917Z
             created_string = parse('19700101T000017Z')
@@ -61,6 +123,19 @@ class EverConverter(object):
             content = note.xpath('content')
             if content:
                 raw_text = content[0].text
+                for res_id, res_val in resources_by_id.iteritems():
+                    if res_val["used"] == 0:
+                        print "unused resource " + res_id
+
+                for res_id, res_val in resources_by_id.iteritems():
+
+                    reference_rx = re.compile('<en-media.+?hash="' + res_id + '"[^>]+?>')
+                    replacement = res_val["tag"].format(res_val["filename"], res_val["filename"])
+                    (raw_text, subs) = re.subn(reference_rx, replacement, raw_text)
+                    if subs == 0:
+                        print raw_text.encode("utf-8")
+                        print "no subs for " + res_id + " " + res_val["filename"]
+
                 # TODO: Option to go to just plain text, no markdown
                 converted_text = self._convert_html_markdown(title, raw_text)
                 if self.fmt == 'csv':
@@ -123,27 +198,29 @@ class EverConverter(object):
             elif not os.path.exists(self.simple_filename):
                 os.makedirs(self.simple_filename)
             for note in notes:
-                # Overwrite duplicates
-                # output_file_path = os.path.join(self.simple_filename, note['title'] + '.md')
-                # Check for duplicates
-                # Filename is truncated to 100 characters (Windows MAX_PATH is 255) - not a perfect fix but an improvement
-                filename = self._format_filename(note['title'][0:99])
-                output_file_path_no_ext_original = os.path.join(self.simple_filename, filename)
-                output_file_path_no_ext = output_file_path_no_ext_original
-                count = 0
-                while os.path.isfile(output_file_path_no_ext + ".md"):
-                    count = count + 1
-                    output_file_path_no_ext = output_file_path_no_ext_original + " (" + str(count) + ")"
-                output_file_path = output_file_path_no_ext + ".md"
+                output_file_path = os.path.join(self.simple_filename, note["filename"]) + ".md"
+                if os.path.exists(output_file_path):
+                    raise Exception("file already exists" + unicode(output_file_path))
                 with io.open(output_file_path, 'w', encoding='utf8') as output_file:
                     if self.metadata:
                         output_file.write(self._metadata(note))
                     output_file.write(note['content'])
+                for resource in note['resources']:
+                    resource_output_path = os.path.join(self.simple_filename, resource["filename"])
+                    if os.path.exists(resource_output_path):
+                        raise Exception("file already exists" + unicode(resource_output_path))
+                    rh = open(resource_output_path, "wb")
+                    rh.write(resource["data"].decode("base64"))
+                    rh.close()
+                with open(output_file_path, 'w') as output_file:
+                    output_file.write(note['content'].encode(encoding='utf-8'))
 
     def _format_filename(self, s):
         for c in r'[]/\;,><&*:%=+@!#^()|?^':
             s = s.replace(c, '-')
-        return s
+        for c in r' ':
+            s = s.replace(c, '_')
+        return unicodedata.normalize('NFKD', unicode(s)).encode('ascii','ignore')
 
     def _metadata(self, note):
         """
